pyproject.toml
[project]
name = "ruralbrain"
version = "0.1.0"
description = "乡村复杂决策系统智慧大脑"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "langchain>=1.1.0",
    "langchain-deepseek>=1.0.0",
    "langchain-zhipuai>=0.0.1",
    "langgraph>=1.0.4",
    "langsmith>=0.4.38",
    "requests>=2.31.0",
    # 害虫检测API依赖
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.4.2",
    "pydantic-settings>=2.0.3",
    "python-multipart>=0.0.6",
    # 图像处理和深度学习依赖
    "flask>=3.1.2",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "torch>=2.0.1",
    "torchvision>=0.15.2",
    "ultralytics>=8.0.0",
    "websocket-client>=1.6.0",
    # RAG 知识库依赖
    "langchain-chroma>=0.1.0",
    "langchain-huggingface>=0.1.0",
    "langchain-text-splitters>=0.3.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "python-pptx>=1.0.0",
    "pypdf>=5.0.0",
    "python-docx>=1.1.0",
    "filetype>=1.2.0",
    "rich>=13.0.0",
    # 测试依赖
    "aiohttp>=3.9.0",
]
